pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "snge"
version = "0.1.0"
description = "Stochastic simulations of Nucleation-Growth-Etching model for graphene synthesis"
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.20",
    "scipy>=1.7",
    "matplotlib>=3.4",
    "tqdm>=4.0",
    "numba>=0.56",
    "pandas>=1.3",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]
gpu = ["cupy-cuda12x"]  # For GPU acceleration, adjust cuda version as needed

[project.scripts]
snge = "snge.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')"
]
filterwarnings = ["ignore::DeprecationWarning"]
addopts = "-v --tb=short"
